fetch restructured with custom hook
